Remove basic spec
This removes the "basic_spec" as it is largely a sanity check against
the env/tooling.  The spec itself is problematic because in Puppet
Enterprise a module will be in 'sitemoduledir' or 'distmoduledir'
depending on whether it's included in PE as part of the distribution or
not.  However it will always be in 'distmoduledir' for FOSS.