Allow to specify an ensure_absent_val
This commit aims to add a new feature for the ini_setting provider, this
feature aims to simulate the ensure => absent behavior when a specific
keyword is specified.

Currently a pattern we have is

```puppet
if $myvar {
  ini_setting { 'SECTION/setting' : value => $myvar }
} else {
  ini_setting { 'SECTION/setting' : ensure => absent }
}
```

If one has dozens or hundreds of parameters to handle then it can easily
make the manifest hard to read.

The solution offer here would turn the above example in something like

```puppet
Ini_setting {
  ensure_absent_val = 'nil' # It is the default
}

ini_setting { 'SECTION/setting' : value => $myvar }
```

If `$myvar` is 'nil' then it will act as if `ensure => absent` would
have been specified. 'nil' being a valid value in some case, this is why
it has been parametrized.

This patch was necessary because none of `value => undef` or `value =>
nil` had the expected behavior.