(MODULES-4932) fix for mimicking commented settings
inifile, by design, looks for a commented version of a setting if an actual one does not exist. if one is found, it attempts to insert the new setting directly below its commented self and mimics the spacing used in the comment. this removes the mimicry and uses the designated key_val_separator instead.