(MODULES-5172) Backwards compatible ini_file.set_value
The fix for MODULES-4932 was not backwards compatible as it changed the
number of required parameters for the set_value function in the ini_file
utility.  This breaks anyone who were also extending the ini_setting
providers and leveraging this function. This change updates the
set_value function to still work with the previous 3 parameters while
supporting the new format with 4 parameters.