MODULES-3624 Allow setting indent character (#237)
* MODULES-3624 Allow setting indent character
Add a new param 'indent' which allows the character used for indentation
of newly-inserted values to be set instead of just using a space. Some
software requires tabs in ini files, for example.
The default is ' ', so the default behaviour is unchanged. This change
only affects settings which are actually created. Settings which already
exist and have a new value will not be re-indented.
A limitation of this approach is that it will use whatever you specify
regardless of which characters the existing indentation uses. That's how
it behaves now with spaces, it might just be less obvious when one is
specifying explicitly.
* Start adding support for specified indent char
* Allow setting of both indent_char and indent_width
* Update existing setting test to check indent_width does nothing
* Allow indent_width to be a string containing an integer