(maint) Fix spec tests on Windows
Previously the spec tests were failing on Windows due to linux paths being used
on a Windows platform.  This commit branches the test to use a correct absolute
path depending on the type of platform that's actually running the test.  This
is required as some functions within Puppet cannot be mocked correctly, such
as path validation.