(MODULES-1821) Handle empty sections
A resource can now have a nil setting and value;
this indicates a section that has no settings,
and hence is empty.

Also, the global setting always exists, even for
empty files.

Finally, this changes how the global section is considered empty:
since the global section has no header, we can't call it empty
when its starting & ending lines are zero.  That, in fact, is
when it includes a single setting or comment.  Instead, we
consider the global section empty whenever it's new.

(Apologies, I'm trying to stage these git commits in a logical way,
but I'm working with changes made in an SVN repo that were much
more haphazard.  Until I stash the changes I have yet to commit,
I won't know whether these committed changes pass the tests.  So,
this commit may break tests.  We'll see.)