(MODULES-1821) Fix regressions in commit 8e72205
Prior to supporting empty sections, when the
the final line in a section was removed by
deleting its sole setting, the section's header
was also removed, thereby deleting the section.

For now, restore that behavior to preserve
backwards compatibility.

There may come a time when, in the world of
empty sections, deleting such a setting should
*not* cause the section header to be deleted.

That's a debate for the future.