Handle backwards compatibility with force_new_section_creation
Prior to the introduction of this new parameter in the ini_setting type,
a user may have had their own custom type which implemented the required
parameters. The introduction of this new type would break anyone who was
not inheriting from the provide ini_setting type. This change adds a
check before looking for this option to allow those users to continue to
work.  If the force_new_section_creation is not defined on the type, the
previous expected functionality of automatic creation of the section
will occur.

Related-Bug: https://bugs.launchpad.net/puppet-openstacklib/+bug/1778247