Finish API conversion of `create_ini_settings`
In https://github.com/puppetlabs/puppetlabs-inifile/pull/373 the
`create_ini_settings` function was ported to the new function API and
namespaced. The boilerplate wasn't fixed up and the function didn't even
work...

This commit finishes the conversion with the following changes:

* Remove the old API `puppet/parser/function` version.
* Create a replacement non-namespaced, (but deprecated), new API shim
that calls the namespaced version.
* Replace argument validation in the function code with data-type based
validation for the individual function parameters.
* Use `call_function` instead of `function_create_resources`.
* Update the README and REFERENCE.md files.
* Test both shim and namespaced version of the function.