match section names containing prefix character
Currently sections with names including the prefix (normally [) can be
created by this provider, but they are then not matched by the regex,
meaning puppet will create a new section (with the same name) every time
it runs. Fix that by allowing the section prefix in the section name.
Any line that starts with the correct prefix (normally [) and ends with
the correct suffix (normally ]) is treated as a valid section header,
regardless of what is between them.

This is needed for splunk inputs.conf file where regexes, including
square brackets, are allowed inside the section name.