(MODULES-11290) Add note about create_ini_settings
When using inifile::create_ini_settings you the resource that is created
will not have the same set of namespace tags as the resource that is
created by the ini_setting resource.

This commit adds a small note to the limitations section to explain
this. Hopefully clearing up any confusion for users of the module.